Move AttributeSet into its own file
